+++ apps/cowswap-frontend/src/common/hooks/useMenuItems.ts
@@ -1,6 +1,7 @@
 import { useMemo } from 'react'
 
 import { useFeatureFlags } from '@cowprotocol/common-hooks'
+import { isLocal } from '@cowprotocol/common-utils'
 
 import { HOOKS_STORE_MENU_ITEM, MENU_ITEMS } from '../constants/routes'
 
@@ -8,6 +9,6 @@ export function useMenuItems() {
   const { isHooksStoreEnabled } = useFeatureFlags()
 
   return useMemo(() => {
-    return isHooksStoreEnabled ? MENU_ITEMS.concat(HOOKS_STORE_MENU_ITEM) : MENU_ITEMS
+    return isHooksStoreEnabled || isLocal ? MENU_ITEMS.concat(HOOKS_STORE_MENU_ITEM) : MENU_ITEMS
   }, [isHooksStoreEnabled])
 }