+++ apps/cowswap-frontend/src/legacy/components/TransactionSettings/index.tsx
@@ -31,7 +31,7 @@ import { useIsEoaEthFlow } from 'modules/swap/hooks/useIsEoaEthFlow'
 import { useIsSlippageModified } from 'modules/swap/hooks/useIsSlippageModified'
 import { useIsSmartSlippageApplied } from 'modules/swap/hooks/useIsSmartSlippageApplied'
 import { useSetSlippage } from 'modules/swap/hooks/useSetSlippage'
-import { useDefaultSwapSlippage, useSwapSlippage } from 'modules/swap/hooks/useSwapSlippage'
+import { useDefaultSwapSlippage, useSmartSwapSlippage, useSwapSlippage } from 'modules/swap/hooks/useSwapSlippage'
 import { getNativeOrderDeadlineTooltip, getNonNativeOrderDeadlineTooltip } from 'modules/swap/pure/Row/RowDeadline'
 import { getNativeSlippageTooltip, getNonNativeSlippageTooltip } from 'modules/swap/pure/Row/RowSlippageContent'
 
@@ -47,7 +47,7 @@ enum DeadlineError {
   InvalidInput = 'InvalidInput',
 }
 
-const Option = styled(FancyButton)<{ active: boolean }>`
+const Option = styled(FancyButton) <{ active: boolean }>`
   margin-right: 8px;
 
   :hover {
@@ -75,7 +75,7 @@ export const Input = styled.input`
   text-align: right;
 `
 
-export const OptionCustom = styled(FancyButton)<{ active?: boolean; warning?: boolean }>`
+export const OptionCustom = styled(FancyButton) <{ active?: boolean; warning?: boolean }>`
   height: 2rem;
   position: relative;
   padding: 0 0.75rem;
@@ -84,7 +84,7 @@ export const OptionCustom = styled(FancyButton)<{ active?: boolean; warning?: bo
 
   :hover {
     border: ${({ theme, active, warning }) =>
-      active && `1px solid ${warning ? darken(theme.error, 0.1) : darken(theme.bg2, 0.1)}`};
+    active && `1px solid ${warning ? darken(theme.error, 0.1) : darken(theme.bg2, 0.1)}`};
   }
 
   input {
@@ -185,6 +185,9 @@ export function TransactionSettings() {
   const defaultSwapSlippage = useDefaultSwapSlippage()
   const setSwapSlippage = useSetSlippage()
   const isSmartSlippageApplied = useIsSmartSlippageApplied()
+  const smartSlippage = useSmartSwapSlippage()
+
+  const chosenSlippageMatchesSmartSlippage = smartSlippage && new Percent(smartSlippage, 10_000).equalTo(swapSlippage)
 
   const [deadline, setDeadline] = useUserTransactionTTL()
 
@@ -256,10 +259,10 @@ export function TransactionSettings() {
         if (
           !Number.isInteger(parsed) || // Check deadline is a number
           parsed <
-            (isEoaEthFlow
-              ? // 10 minute low threshold for eth flow
-                MINIMUM_ETH_FLOW_DEADLINE_SECONDS
-              : MINIMUM_ORDER_VALID_TO_TIME_SECONDS) || // Check deadline is not too small
+          (isEoaEthFlow
+            ? // 10 minute low threshold for eth flow
+            MINIMUM_ETH_FLOW_DEADLINE_SECONDS
+            : MINIMUM_ORDER_VALID_TO_TIME_SECONDS) || // Check deadline is not too small
           parsed > MAX_DEADLINE_MINUTES * 60 // Check deadline is not too big
         ) {
           setDeadlineError(DeadlineError.InvalidInput)
@@ -318,7 +321,7 @@ export function TransactionSettings() {
             </Option>
             <OptionCustom active={isSlippageModified} warning={!!slippageError} tabIndex={-1}>
               <RowBetween>
-                {!isSmartSlippageApplied && (tooLow || tooHigh) ? (
+                {!isSmartSlippageApplied && !chosenSlippageMatchesSmartSlippage && (tooLow || tooHigh) ? (
                   <SlippageEmojiContainer>
                     <span role="img" aria-label="warning">
                       ⚠️
@@ -336,7 +339,7 @@ export function TransactionSettings() {
               </RowBetween>
             </OptionCustom>
           </RowBetween>
-          {!isSmartSlippageApplied && (slippageError || tooLow || tooHigh) ? (
+          {!isSmartSlippageApplied && !chosenSlippageMatchesSmartSlippage && (slippageError || tooLow || tooHigh) ? (
             <RowBetween
               style={{
                 fontSize: '14px',
@@ -398,8 +401,8 @@ export function TransactionSettings() {
                     deadlineInput.length > 0
                       ? deadlineInput
                       : deadline === DEFAULT_DEADLINE_FROM_NOW
-                      ? ''
-                      : (deadline / 60).toString()
+                        ? ''
+                        : (deadline / 60).toString()
                   }
                   onChange={(e) => parseCustomDeadline(e.target.value)}
                   onBlur={() => {